[José Llancamil] Bug Fix y más pruebas unitarias

+++ repair-management-system/src/main/java/com/autofix/repairmanagementsystem/dto/RepairTypeMotorSummaryDTO.java
@@ -3,13 +3,11 @@ package com.autofix.repairmanagementsystem.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 public class RepairTypeMotorSummaryDTO {
     private String repairTypeDescription;
     private String engineType;
     private Long vehicleCount;
-    private BigDecimal totalCost;
+    private Long totalCost;
 }
